+++ src/rpc/mnemonic.cpp
@@ -220,7 +220,7 @@ UniValue mnemonic(const JSONRPCRequest &request)
             // m / purpose' / coin_type' / account' / change / address_index
             CExtKey ekDerived;
             ekMaster.Derive(ekDerived, BIP44_PURPOSE);
-            ekDerived.Derive(ekDerived, Params().BIP44ID());
+            ekDerived.Derive(ekDerived, (uint32_t)Params().BIP44ID());
 
             eKey58.SetKey(CExtKeyPair(ekDerived), CChainParams::EXT_SECRET_KEY);
             result.pushKV("derived", eKey58.ToString());
